+++ src/test/java/com/ojins/chatbot/testService.java
@@ -1,8 +1,8 @@
 package com.ojins.chatbot;
 
-import com.ojins.chatbot.dialog.StateIO;
 import com.ojins.chatbot.service.QAService;
 import com.ojins.chatbot.service.QAServiceBuilder;
+import com.ojins.chatbot.util.StateIO;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.junit.Assert;
